refactor: centralize datetime helpers
